Add a property-based test
Here we add a dependency on hypothesis, a property-test framework that integrates with pytest.
I've added one test that creates random operons and checks whether all features have the same
orientation. We don't actually look at the output of the test, but it will only pass if there
are no uncaught exceptions. Essentially, I'm just fuzzing the code.