Resovles #37.
In normal use we'll never have just a single pipeline output file, and in fact will probably have thousands.
We can concatenate all of the results together and stream them into analyze(), but we have to check each line
to determine whether or not its a header line. This commit makes this change and adds a few tests.
It also adds a header line to one of the test CSVs which was inappropriately missing.