Resolves #41.
Certain proteins were not being plotted, seemingly at random, including proteins that were being used as seeds.
The issue was that the very first record was erroneously being skipped since we had previously assumed that there
would always be a CSV header. Due to the changes made in #37, deleting the code that skips the first line does not
actually affect any existing tests. Regression tests are added here to ensure the presence or absense of a header
is handled correctly and that no proteins go missing.