Fix distance finding when multiple copies of the same feature exist
We were just giving up in the event where two or more features had the same name and this led to a situation where a distance was not being correctly calculated.
This commit changes our strategy and requires that any rule that applies to a particular feature must pass for all copies.

Resolves #51.