Add some tools to compare pipeline results
After running the pipeline using curated databases, we'll typically want
to re-BLAST using a more general database to see if a better match can
be found. This is important since in practice, we're using very low
e-values. These functions take two list of operons, find features that
occupy the same regions of the same contig, and see how they compare.

There is a single function that lets the user determine if a certain
fraction of reannotations meet a given regex, which allows them to pose
questions like "is there at least one transposase that was reannotated
as either "transposase", "integrase", or "recombinase" more than 50% of
the time?" We may want to add more functions in the future, but that
will be part of a different issue.

There are no tests in this commit, and those are pretty critical though
everything appears to be working as expected.