Fully implement nucleotide BLAST searches
This commit adds all the functionality needed to run nucleotide BLAST in the
pipeline. It works using real data searching for tRNAs, but I still need
to add a few unit and/or integration tests.

In general, there's a lot of copying and pasting from the
blastp-handling code, but I think it's not worth it to make the existing
functions more general since there would be special casing all over the
place, and barring the discovery of alien life, we'll never need to
search for anything other than DNA and proteins.