Add operon sequence methods
I'm going to be looking at ways to deduplicate highly-similar clusters.
In order to do that I need to be able to get the raw DNA sequence that
the features of interest bound. I'm deliberately not adding any method
to load the raw contig sequence data as that may vary.