Deduplication of operons
Adds function to approximately remove operons that are roughly identical
within the bounds of the identified features. This can result in false
positives - that is, operons that in reality are unique but are flagged
as redundant, however, the method is quite fast and in reality, such
false positives probably vary by only a few nucleotides.