Self-targeting spacer search
This commit provides functionality needed to identify locations in a contig that are targeted by a spacer. It needs testing and documentation. Additionally, the user specifies the threshold for what they consider to be self-targeting, however, this should probably be a ratio of matches to spacer length, rather than just raw matches.