Improve spacer search
This should fix most bugs, and also parallelizes the self-targeting
spacer search. It seems to be fast enough now that we can do a brute
force search on the entire dataset using our local machine.

I think there are still a few gains to be had, since there are a few
hacks in here which were needed to work around some bugs in the
pairwise alignment library, but those aren't optimal. We also need
more documentation and tests.