Reserialize operons
This commit allows operons to be re-written as strings in a form that's
effectively identical to how they appeared in the original CSV
(meaningless differences do occur when floats have trailing zeros or
things like that).

The purpose of this is to allow a large speedup of a pipeline. When
evaluating rules, the operons must be loaded from disk and parsed, evaluated againt the rules, and then the entire set of operons must be loaded and parsed again and the
passing operons written out (alternatively, the operons can be
serialized with pickle and saved as a file). In both cases, an elegant
chain of pipes is not possible, and I'd prefer text to binary pickle
files as it will future-proof data against changes to the library.